scripts/requirements-EJ.txt
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3
google-generativeai==0.3.2
PyPDF2==3.0.1
pillow==10.0.1
PyMuPDF>=1.23.0

# Configure config.py com:
# - Sua GEMINI_API_KEY

# - A URL do PDF de empresas juniors

# - Para iniciar o processo após a configuração, ative extrair_empresas_juniores.py 
